реализовано создание и вывод рекламаций в сервисе Garanties.

+++ GateWay/src/main/java/ru/sem/gateway/garanties/GarantRequestDto.java
@@ -18,7 +18,7 @@ public class GarantRequestDto {
 
     String clientName;
 
-    Integer clientPhone;
+    Long clientPhone;
 
     String transportModel;
